[python] Lower the constraint bound on NumPy runtime dependency
Using the build time NumPy version as a lower bound for the Python
dependency range is too much of a constraint, in particular when
trying to install gstlearn with Python packages that need older but
compatible versions of NumPy, such as tensorflow.

Indeed, building gstlearn with NumPy>=2.0 makes the Python package
still compatible with NumPy
1 (c.f. https://numpy.org/doc/2.1/dev/depending_on_numpy.html#numpy-2-0-specific-advice).

Instead of using the NumPy version from build time, rather use the
latest supported version of NumPy as a lower bound for the run-time
Python package dependency range.

As of now, NumPy 1.24 is the oldest supported version (to be dropped
at the end of 2024, c.f. https://numpy.org/neps/nep-0029-deprecation_policy.html).