Add a preprocessor variable GSTLEARN_VERSION_NUMBER in version.h.
The variable GSTLEARN_VERSION cannot really be used for anything else than displaying it because it's a string (so we cannot do e.g. #if GSTLEARN_VERSION == "1.5.0"). We could pass the individual bits of the version as defined by CMake (major, minor, patch), but my experience is that it is convenient to wrap them in a single number.

It is built as MAJOR*10000 + 100*MINOR + PATCH which leaves space for up to 99 minor/patches, so in practice it usually fine.