Fix previous commit that added USE_BOOST_SPAN when requested as this was not passed to projects using gstlearn.
Projects that include gstlearn need to know that USE_BOOST_SPAN was defined when gstlearn was compiled, which means the #define must be exported. That requires setting it as PUBLIC, which cannot be done with add_definitions() (or add_compile_definitions()). Instead, we must use target_compile_definitions() which has to be set on each target (that get exported), so this is a tiny bit less convenient.

This commit fixes the python_build target (in addition to the C++ libraries), it is possible that the equivalent target for R also needs to be fixed, I haven't checked.