Issue #500: refactor SPDEOp and related classes to make it easy to change the solver algorithm (currently Eigen::conjugate_gradient()).
* ALinearOpEigenCG now derives virtually from ALinearOp, to allow e.g., SPDEOp to derive from this class and another one (ASPDEOp) that implements some virtual functions of ALinearOp.
* The LinearOpCGSolver<TLinOP> now derives from a (non template) base class ALinearOpCGSolver, which allows other classes (e.g., ASPDEOp) to store a pointer to the base class.
* There is a new class ASPDEOp which contains all of the code previously in SPDEOp, and stores a pointer to a ALinearOpCGSolver instead of an actual LinearOpCGSolver<TLinOP>. This class derives virtually from ALinearOp to provide an implementation of the required virtual functions.
* SPDEOp now derives from ASPDEOp and contains essentially one line, which is allocating the ALinearOpCGSolver in the base class to a LinearOpCGSolver<SPDEOp>.

To change the solver, add a new class similar to SPDEOp, and add a specialization of Eigen::internal::conjugate_gradient() that uses this new type. There is a (commented out) class ExampleSPDEOp at the end of SPDEOp.hpp that shows how to do that.

The refactoring allows making SPDEOp tiny so that reimplementing a new version of it is trivial. The other changes are required to ensure the proper final template type is used (e.g., that the LinearOpCGSolver<> uses SPDEOp and not ASPDEOp).