Fix a Windows installation bug, the .lib file needed to link with the DLL was not installed.
This is because the install() command specified that it wants the components RUNTIME (.dll) and ARCHIVE (.lib), but did not specify a DESTINATION folder for ARCHIVE since the default (lib/) is what we want. However CMake's doc says:

"For shared libraries on DLL platforms, if [...] either a RUNTIME or ARCHIVE destination is specified, the component is installed to that destination, and the other component is not installed."

Note that it doesn't say "if either RUNTIME or ARCHIVE is specified" but "if either a RUNTIME or ARCHIVE **destination** is specified," meaning that specifying ARCHIVE (in our case) without giving it a destination causes it to be skipped. This doesn't make sense to me, but this is what the doc says (and the code does), so now the destination is specified again.